+++ imputena/multiple_imputation/mice.py
@@ -6,7 +6,7 @@ from random import shuffle
 from imputena import mean_substitution, linear_regression, logistic_regression
 
 
-def mice(data=None, imputations=3, regressions='available'):
+def mice(data=None, imputations=3):
     """Performs multiple imputation by chained equations (MICE) on the data.
     Several (parameter imputations) linear regression imputations are
     performed on the dataset. For each one, the a random order of imputation of
@@ -21,12 +21,6 @@ def mice(data=None, imputations=3, regressions='available'):
     :type data: pandas.DataFrame
     :param imputations: Number of imputations to perform
     :type imputations: scalar, default 3
-    :param regressions: If 'available': Impute missing values by modeling a
-        regression based on all available predictors if some predictors have
-        missing values themselves. If 'complete': Only impute with a
-        regression model based on all predictors and leave missing values in
-        rows in which some predictor value is missing itself unimputed.
-    :type regressions: {'available', 'complete'}, default 'available'
     :return: A list of MICE imputations performed with randomly chosen
         orders of column imputations.
     :rtype: list of pandas.DataFrame
@@ -35,31 +29,21 @@ def mice(data=None, imputations=3, regressions='available'):
     # Check if data is a dataframe:
     if not isinstance(data, pd.DataFrame):
         raise TypeError('The data has to be a DataFrame.')
-    # Check that the value of regressions is valid:
-    if regressions not in ['available', 'complete']:
-        raise ValueError(regressions + 'could not be understood.')
     # Create the list that will be returned
     imputed_datasets = []
     # Impute several times and add the results to the list:
     for _ in range(imputations):
-        imputed_datasets.append(
-            mice_one_imputation(data, regressions))
+        imputed_datasets.append(mice_one_imputation(data))
     # Return the imputed datasets:
     return imputed_datasets
 
 
-def mice_one_imputation(data, regressions):
+def mice_one_imputation(data):
     """Auxiliary function that performs one MICE imputation, choosing the
     order in which the columns are imputed at random.
 
     :param data: The data on which to perform the imputation.
     :type data: pandas.DataFrame
-    :param regressions: If 'available': Impute missing values by modeling a
-        regression based on all available predictors if some predictors have
-        missing values themselves. If 'complete': Only impute with a
-        regression model based on all predictors and leave missing values in
-        rows in which some predictor value is missing itself unimputed.
-    :type regressions: {'available', 'complete'}
     :return: The dataframe with one MICE imputation performed.
     :rtype: pandas.DataFrame
     """
@@ -80,11 +64,9 @@ def mice_one_imputation(data, regressions):
     for column in columns_with_na:
         if is_numeric_dtype(data[column]):
             res.loc[na_mask[column], column] = np.nan
-            linear_regression(
-                res, column, regressions=regressions, inplace=True)
+            linear_regression(res, column, inplace=True)
         else:
             res.loc[na_mask[column], column] = None
-            logistic_regression(
-                res, column, regressions=regressions, inplace=True)
+            logistic_regression(res, column, inplace=True)
     return res